IGNITE-18156 Simplification, test suite split.
